+++ app/src/main/java/com/thorn/wego/Login/Model/LoginUser.java
@@ -43,7 +43,7 @@ public class LoginUser implements ILoginUser {
     public String getLoginCode(){ return basicNetworkJson.getCode();}
 
     @Override
-    public int getUserId(){return basicNetworkJson.getUser().getUserid();}
+    public long getUserId(){return basicNetworkJson.getUser().getUserid();}
 
     @Override
     public int checkUser(String url){
